step importer: use double precision

+++ src/util/step_importer.cpp
@@ -55,7 +55,7 @@ struct DATA;
 
 static bool processNode(const TopoDS_Shape &shape, DATA &data);
 
-static bool processComp(const TopoDS_Shape &shape, DATA &data, const glm::mat4 &mat_in = glm::mat4(1));
+static bool processComp(const TopoDS_Shape &shape, DATA &data, const glm::dmat4 &mat_in = glm::dmat4(1));
 
 struct DATA {
     Handle(TDocStd_Document) m_doc;
@@ -119,7 +119,8 @@ static bool getColor(DATA &data, TDF_Label label, Quantity_Color &color)
     return false;
 }
 
-static bool processFace(const TopoDS_Face &face, DATA &data, Quantity_Color *color, const glm::mat4 &mat = glm::mat4(1))
+static bool processFace(const TopoDS_Face &face, DATA &data, Quantity_Color *color,
+                        const glm::dmat4 &mat = glm::dmat4(1))
 {
     if (Standard_True == face.IsNull())
         return false;
@@ -222,7 +223,7 @@ static bool processFace(const TopoDS_Face &face, DATA &data, Quantity_Color *col
 }
 
 static bool processShell(const TopoDS_Shape &shape, DATA &data, Quantity_Color *color,
-                         const glm::mat4 &mat = glm::mat4(1))
+                         const glm::dmat4 &mat = glm::dmat4(1))
 {
     TopoDS_Iterator it;
     bool ret = false;
@@ -237,7 +238,7 @@ static bool processShell(const TopoDS_Shape &shape, DATA &data, Quantity_Color *
     return ret;
 }
 
-static bool processSolid(const TopoDS_Shape &shape, DATA &data, const glm::mat4 &mat_in = glm::mat4(1))
+static bool processSolid(const TopoDS_Shape &shape, DATA &data, const glm::dmat4 &mat_in = glm::dmat4(1))
 {
     TDF_Label label = data.m_assy->FindShape(shape, Standard_False);
     bool ret = false;
@@ -259,14 +260,14 @@ static bool processSolid(const TopoDS_Shape &shape, DATA &data, const glm::mat4 
     gp_Trsf T = loc.Transformation();
     gp_XYZ coord = T.TranslationPart();
 
-    auto mat = mat_in * glm::translate(glm::vec3(coord.X(), coord.Y(), coord.Z()));
+    auto mat = mat_in * glm::translate(glm::dvec3(coord.X(), coord.Y(), coord.Z()));
 
     gp_XYZ axis;
     Standard_Real angle;
 
     if (T.GetRotation(axis, angle)) {
-        glm::vec3 gaxis(axis.X(), axis.Y(), axis.Z());
-        float angle_f = angle;
+        glm::dvec3 gaxis(axis.X(), axis.Y(), axis.Z());
+        double angle_f = angle;
         mat = glm::rotate(mat, angle_f, gaxis);
     }
 
@@ -281,21 +282,21 @@ static bool processSolid(const TopoDS_Shape &shape, DATA &data, const glm::mat4 
     return ret;
 }
 
-static bool processComp(const TopoDS_Shape &shape, DATA &data, const glm::mat4 &mat_in)
+static bool processComp(const TopoDS_Shape &shape, DATA &data, const glm::dmat4 &mat_in)
 {
     TopoDS_Iterator it;
     TopLoc_Location loc = shape.Location();
     gp_Trsf T = loc.Transformation();
     gp_XYZ coord = T.TranslationPart();
 
-    auto mat = mat_in * glm::translate(glm::vec3(coord.X(), coord.Y(), coord.Z()));
+    auto mat = mat_in * glm::translate(glm::dvec3(coord.X(), coord.Y(), coord.Z()));
 
     gp_XYZ axis;
     Standard_Real angle;
 
     if (T.GetRotation(axis, angle)) {
-        glm::vec3 gaxis(axis.X(), axis.Y(), axis.Z());
-        float angle_f = angle;
+        glm::dvec3 gaxis(axis.X(), axis.Y(), axis.Z());
+        double angle_f = angle;
         mat = glm::rotate(mat, angle_f, gaxis);
     }
